+++ dqops/src/main/frontend/src/components/DefinitionLayout/DefinitionTree.tsx
@@ -25,7 +25,7 @@ import DataQualityContextMenu from './DataQualityContextMenu';
 import RuleContextMenu from './RuleContextMenu';
 import SensorContextMenu from './SensorContextMenu';
 
-const defaultChecks = ['Table-level policies', 'Column-level policies'];
+const defaultChecks = ['Table-level quality policies', 'Column-level quality policies'];
 
 export const DefinitionTree = () => {
   const dispatch = useActionDispatch();