+++ dqops/src/main/java/com/dqops/rest/controllers/TableQualityPoliciesController.java
@@ -118,7 +118,7 @@ public class TableQualityPoliciesController {
             List<TableQualityPolicyListModel> models = patternWrappersList.stream()
                     .map(pw -> TableQualityPolicyListModel.fromPatternSpecification(pw.getSpec(), canEdit))
                     .collect(Collectors.toList());
-            models.sort(Comparator.comparing(model -> model.getPatternName()));
+            models.sort(Comparator.comparing(model -> model.getPolicyName()));
 
             return new ResponseEntity<>(Flux.fromStream(models.stream()), HttpStatus.OK);
         }));
@@ -205,8 +205,8 @@ public class TableQualityPoliciesController {
 
             boolean canEdit = principal.hasPrivilege(DqoPermissionGrantedAuthorities.EDIT);
             TableQualityPolicyModel patternModel = new TableQualityPolicyModel() {{
-                setPatternName(patternName);
-                setPatternSpec(defaultChecksPatternWrapper.getSpec());
+                setPolicyName(patternName);
+                setPolicySpec(defaultChecksPatternWrapper.getSpec());
                 setCanEdit(canEdit);
                 setYamlParsingError(defaultChecksPatternWrapper.getSpec().getYamlParsingError());
             }};
@@ -296,7 +296,7 @@ public class TableQualityPoliciesController {
             @ApiParam("Pattern name") @PathVariable String patternName,
             @ApiParam("Default checks pattern model") @RequestBody TableQualityPolicyModel patternModel) {
         return Mono.fromFuture(CompletableFuture.supplyAsync(() -> {
-            if (patternModel == null || Strings.isNullOrEmpty(patternName) || patternModel.getPatternSpec() == null) {
+            if (patternModel == null || Strings.isNullOrEmpty(patternName) || patternModel.getPolicySpec() == null) {
                 return new ResponseEntity<>(Mono.empty(), HttpStatus.NOT_ACCEPTABLE);
             }
 
@@ -313,7 +313,7 @@ public class TableQualityPoliciesController {
                         }
 
                         TableQualityPolicyWrapper defaultChecksPatternWrapper = defaultChecksPatternsList.createAndAddNew(patternName);
-                        defaultChecksPatternWrapper.setSpec(patternModel.getPatternSpec());
+                        defaultChecksPatternWrapper.setSpec(patternModel.getPolicySpec());
                         userHomeContext.flush();
 
                         return new ResponseEntity<>(Mono.empty(), HttpStatus.CREATED);
@@ -466,7 +466,7 @@ public class TableQualityPoliciesController {
             @ApiParam("Pattern name") @PathVariable String patternName) {
         return Mono.fromFuture(CompletableFuture.supplyAsync(() -> {
 
-            if (Strings.isNullOrEmpty(patternName) || patternModel == null || patternModel.getPatternSpec() == null) {
+            if (Strings.isNullOrEmpty(patternName) || patternModel == null || patternModel.getPolicySpec() == null) {
                 return new ResponseEntity<>(Mono.empty(), HttpStatus.NOT_ACCEPTABLE);
             }
 
@@ -480,10 +480,10 @@ public class TableQualityPoliciesController {
 
                         if (existingDefaultChecksPatternWrapper == null) {
                             TableQualityPolicyWrapper defaultChecksPatternWrapper = defaultChecksPatternsList.createAndAddNew(patternName);
-                            defaultChecksPatternWrapper.setSpec(patternModel.getPatternSpec());
+                            defaultChecksPatternWrapper.setSpec(patternModel.getPolicySpec());
                         } else {
                             TableQualityPolicySpec oldPatternSpec = existingDefaultChecksPatternWrapper.getSpec(); // just to load
-                            existingDefaultChecksPatternWrapper.setSpec(patternModel.getPatternSpec());
+                            existingDefaultChecksPatternWrapper.setSpec(patternModel.getPolicySpec());
                         }
                         userHomeContext.flush();