+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 * Duplicate record detection check
 * Some column-level checks fixed to skip NULL values, instead of treating them as invalid (not passing checks)
 * Store data lineage definition (links to source tables)
+* Default observability check patterns renamed to data quality policies
+* Default quality policy named "default" deleted, and replaced with multiple smaller policies to allow more control